requirements.txt
fastapi
uvicorn
faiss-cpu
python-multipart
sentence-transformers
assemblyai
google-generativeai
PyPDF2
python-pptx
